Fukc this
Ok I don't have time to write an essay to complain about stupid Vscoide
and it's everything. Why is tasks so garbage??? why is confiruging
npm.packageManager=pnpm not working and it's still insisting on starting
npm for the "npm: watch" task??? isn't it just supposed to automatically
work????? what's the point of this switch if the "npm" tasks don't
change

Ok on the SECOND fucking layer why is after fixign eerythign then the
extension host not debugging breakpoints. Wghy do I have declare this
"referecnes" section in tsconfig.json??? why can't everyhing just be a
single source of truth?????? like people complain about having to list
files in cmake but this is literally worse